Commit 07/11

+++ Estacionamiento/Clases/Vehiculo.php
@@ -63,7 +63,7 @@ class Vehiculo
 		$this->_marca = $marca;
     }
 
-    #FUNCIONES DB --------------------------------------------------------------------
+    #FUNCIONES DB ---------------------------------------------------------------------------------------------------
 
     #INSERTAR VEHICULO EN DB
     
@@ -74,10 +74,10 @@ class Vehiculo
 		{
             $objetoAccesoDato = AccesoDatos::dameUnObjetoAcceso();
             $consulta = $objetoAccesoDato->RetornarConsulta("INSERT into Vehiculo (id_cochera,patente,marca,color,hora,fecha) values (:idCochera,:patente,:marca,:color,:hora,:fecha)"); 			
-			$consulta->bindValue('idCochera',$vehiculo->_idCochera, PDO::PARAM_STR); 
+			$consulta->bindValue(':idCochera',$vehiculo->_idCochera, PDO::PARAM_STR); 
 			$consulta->bindValue(':patente',$vehiculo->_patente, PDO::PARAM_STR);
-			$consulta->bindValue('marca',$vehiculo->_marca, PDO::PARAM_STR);
-			$consulta->bindValue('color',$vehiculo->_color, PDO::PARAM_STR);
+			$consulta->bindValue(':marca',$vehiculo->_marca, PDO::PARAM_STR);
+			$consulta->bindValue(':color',$vehiculo->_color, PDO::PARAM_STR);
 			$consulta->bindValue(':fecha',$vehiculo->_fecha, PDO::PARAM_STR);
 			$consulta->bindValue(':hora',$vehiculo->_hora, PDO::PARAM_STR);
 			$consulta->execute();
@@ -93,32 +93,86 @@ class Vehiculo
 		return $resultado;		
 	}
 
-	#ELIMINAR VEHICULO EN DB
+	#ELIMINAR VEHICULO EN DB ----------------------------------------------------------------------------------------
 	
 	public static function RetirarAuto($patente)
 	{
-		$objetoAccesoDato = AccesoDatos::dameUnObjetoAcceso();
-		$consulta = $objetoAccesoDato->RetornarConsulta("SELECT * FROM Vehiculo WHERE patente = :patente");			
-		$consulta->execute();
-		$resultado = $consulta->rowCount();
+		$patente = Vehiculo::ValidarPatente($vehiculo->GetPatente());
+		if( $patente != FALSE)
+		{
+			$objetoAccesoDato = AccesoDatos::dameUnObjetoAcceso();
+			$consulta = $objetoAccesoDato->RetornarConsulta("SELECT * FROM Vehiculo WHERE patente = :patente");	
+			$consulta->bindValue(':patente',$patente);		
+			$consulta->execute();
+			$resultado = $consulta->rowCount();
 
-		$consulta = $objetoAccesoDato->RetornarConsulta("DELETE FROM Vehiculo WHERE patente = :patente");
+			$consulta = $objetoAccesoDato->RetornarConsulta("DELETE FROM Vehiculo WHERE patente = :patente");
+			$consulta->bindValue(':patente',$patente);
+			$consulta->execute();
+
+			$resultado = $resultado - $consulta->rowCount();
+		
+			if($resultado != 0)
+			{
+				$cochera = Cochera::LiberarLugar(ObtenerIdCochera($patente));
+				return "Cantidad filas modificadas = ".$resultado."\n.Cochera".$cochera." liberada.";
+			}
+		}
+		else
+		{
+			$resultado = "Patente no valida";
+		}
+	}
+
+	public static function ObtenerIdCochera($patente)
+	{
+		$cochera = "Error";
+		$objetoAccesoDato = AccesoDatos::dameUnObjetoAcceso();
+		$consulta = $objetoAccesoDato->RetornarConsulta("SELECT * FROM Vehiculo WHERE patente = :patente");
 		$consulta->bindValue(':patente',$patente);
 		$consulta->execute();
 
-		$resultado = $resultado - $consulta->rowCount();
-		
-		if($resultado != 0)
+		while($linea = $consulta->fetch(PDO::FETCH_ASSOC))
 		{
-			return "Cantidad filas modificadas = ".$resultado;
+			$cochera = $linea["id_cochera"];
 		}
-
-		//Cochera::LiberarLugar($id);
+		return $cochera;
 	}
 
-    
 
+	# CALCULA MONTO DE COBRO POR HORAS TRANSCURRIDAS --------------------------------------------------------------------
+	
+	public static function CalcularMonto($tiempo)
+	{
+		$actual = time();
+		$tiempo = $actual - $tiempo;
+		$horas = round(($tiempo/60)/60,2);
 
+		switch ($horas) 
+		{
+			case ($horas <= 9):
+				$valor = $horas*10;
+				break;
+			case ($horas > 9 && $horas < 12):
+				$valor = 90;
+				break;
+			case ($horas >12 && $horas <16):
+				$diferencia = $horas -12;
+				$valor = 90 + ($diferencia * 10);
+				break;
+			case ($horas > 16 && $horas <= 24):
+				$valor = 170;
+				break;
+			case ($horas > 24):
+				$diferencia = $horas -24;
+				$valor = 170 + ($diferencia * 10);
+				break;
+			default:
+				break;
+		}
+		
+		return $valor;
+	}
 
 
 }
